+++ frontend/web/admin/src/components/CustomerServiceDashboard/ServiceStatistics.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import Table from '../Table';
+import Table from '../Tablecontainer';
 import styles from '../style/ServiceStatistics.module.css';
 
 interface ServiceStatistic {
