+++ frontend/web/admin/src/components/MasterSuperAdminDashboard/AdminActivityLog.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import Table from '../Table';
+import Table from '../Tablecontainer';
 import styles from '../style/AdminActivityLog.module.css';
 
 interface ActivityLogItem {
