+++ frontend/web/admin/src/layouts/MasterSuperAdminDashboard/GlobalFinance.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { MetricCard } from '../../../../components/MetricCard';
-import { MetricChart } from '../../../../components/MetricChart';
-import { Table } from '../../../../components/Table';
+import { MetricCard } from '../../components/MetricCard';
+import { MetricChart } from '../../components/MetricChart';
+import { Table } from '../../components/Table';
 import styles from './GlobalFinance.module.css';
 
 export const GlobalFinance: React.FC = () => {
