+++ frontend/web/admin/src/layouts/MasterSuperAdminDashboard/Overview.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { AppDispatch, RootState } from '../../redux/store';
 import { MetricCard } from '../../components/MetricCard';
-import { Table } from '../../components/Table';
+import { Table } from '../../components/Tablecontainer';
 import { MetricChart } from '../../components/MetricChart';
 import { fetchDashboardMetrics } from '../../redux/slices/dashboardSlice';
 import styles from './Overview.module.css';
